fix: 그룹 내에서 하단 네비게이션 지도 클릭시 에러페이지 뜨는 문제(#187)
- 그룹에서의 지도 nav item 적용이 안된 문제
- 일관성을 위해 /map 경로의 컴포넌트처럼 클라이언트 컴포넌트로 구현
- 같은 로직인 /map 코드를 재사용하기 위해 컴포넌트로 분리
